Restructuring imports fix
