+++ src/features/coal/AssemblyDrillsPage.jsx
@@ -20,8 +20,8 @@ import { useTheme } from "../../shared/context/ThemeContext";
 import usePointerGlow from "../../shared/hooks/usePointerGlow";
 import { coalCourseMeta } from "../../shared/data/coalCourseOutline";
 import "./AssemblyDrillsPage.css";
-import "../Home/Home.css";
-import "../LearningResources/LearningResourcesPage.css";
+import "../home/Home.css";
+import "../learning-resources/LearningResourcesPage.css";
 
 const COAL_ACCENT = coalCourseMeta.accent;
