+++ src/features/coal/CoalHomePage.jsx
@@ -14,8 +14,8 @@ import { useTheme } from "../../shared/context/ThemeContext";
 import usePointerGlow from "../../shared/hooks/usePointerGlow";
 import { coalCourseParts, coalCourseMeta } from "../../shared/data/coalCourseOutline";
 import { getAllCoalModules } from "../../shared/utils/coalCourseUtils";
-import "../Home/Home.css";
-import "../LearningResources/LearningResourcesPage.css";
+import "../home/Home.css";
+import "../learning-resources/LearningResourcesPage.css";
 
 const COAL_ACCENT = coalCourseMeta.accent;
