+++ src/features/coal/InstructionTraceLabPage.jsx
@@ -7,8 +7,8 @@ import { useTheme } from "../../shared/context/ThemeContext";
 import usePointerGlow from "../../shared/hooks/usePointerGlow";
 import { coalCourseMeta } from "../../shared/data/coalCourseOutline";
 import CPUExplorer from "./CPUExplorer";
-import "../Home/Home.css";
-import "../LearningResources/LearningResourcesPage.css";
+import "../home/Home.css";
+import "../learning-resources/LearningResourcesPage.css";
 
 const COAL_ACCENT = coalCourseMeta.accent;
