+++ src/shared/components/topics/PremiumLearningShell.jsx
@@ -4,7 +4,7 @@ import { Home } from "lucide-react";
 import { useTheme } from "../../context/ThemeContext";
 import { useAuth } from "../../../auth/context/AuthContext";
 import progressService from "../../services/progressService";
-import "../../pages/ArithmeticFunctionsAndHDLs/AFHDLLayout.css";
+import "../../../features/arithmetic-hdl/AFHDLLayout.css";
 import "./PremiumLearningShell.css";
 import RelatedSeoLinks from "../../seo/RelatedSeoLinks";
